% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.R
\name{glance.mdl_icomb_lst}
\alias{glance.mdl_icomb_lst}
\title{Glance a mable which contains an information combination reconciliation method}
\usage{
\method{glance}{mdl_icomb_lst}(x, ...)
}
\arguments{
\item{x}{A mable}

\item{...}{Arguments for model methods}
}
\value{
The tibble contains the output of \code{glance()} for that model,
with an added logical column \code{.included} indicating whether the corresponding node is present
in the reconciliation process when the information combination method is used.
}
\description{
It uses the models within a mable to produce a one row summary statistics of their fits.
}
\examples{
\donttest{
library(fable)
library(fabletools)
library(tsibble)
library(dplyr)

tourism_gts <- tourism |>
  aggregate_key(State * Purpose,
                Trips = sum(Trips))

fit <- tourism_gts |>
  model(base = ETS(Trips)) |>
  reconcile(ols = min_trace(base, method = "ols"),
            icomb = icomb(base, train_size = 75))

fit |>
  glance()
}

}
