% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconciliation.R
\name{icomb}
\alias{icomb}
\title{Information combination reconciliation}
\usage{
icomb(
  models,
  train_size,
  alpha = 1,
  standardize = FALSE,
  standardize_response = FALSE,
  intercept = TRUE,
  lambda = NULL,
  lambda_min_ratio = "expand",
  nlambda = 100,
  maxit = 1e+07,
  exact = TRUE
)
}
\arguments{
\item{models}{A column of models in a mable.}

\item{train_size}{The size of the initial training window.}

\item{alpha}{The elasticnet mixing parameter, with \eqn{0 \leq \alpha \leq 1}. The penalty is defined as
\deqn{(1 - \alpha)/2\|B\|_2^2 + \alpha\|B\|_2}
\code{alpha = 1} is the group lasso penalty, and \code{alpha = 0} is the ridge penalty.}

\item{standardize}{Logical flag for \code{x} variable standardization, prior to fitting the model sequence.
The coefficients are always returned on the original scale. Default is \code{standardize = FALSE}.}

\item{standardize_response}{This allows the user to standardize the response variables.
Default is \code{standardize_response = FALSE}.}

\item{intercept}{Should intercepts be fitted (default = TRUE) or set to zero (FALSE).}

\item{lambda}{A user supplied \code{lambda} sequence. Typical usage is to have the program compute its own
\code{lambda} sequence based on \code{nlambda} and \code{lambda_min_ratio}. Supplying a value of \code{lambda} overrides this. Supply a decreasing
sequence of \code{lambda} values.}

\item{lambda_min_ratio}{The smallest value for \code{lambda}, as a fraction of \code{lambda_max}
(the data derived value for which all coefficients are zero). \code{lambda_min_ratio = "expand"} (default) sets the ratio as
\eqn{10^{-\lfloor \log_{10}(\lambda_{max})\rfloor-2}} whereas \code{lambda_min_ratio = "glmnet"} sets the ratio to the value used in the \code{glmnet} package.
If \code{nobs < nvars}, the default is \code{0.01}, otherwise \code{1e-04}.}

\item{nlambda}{The number of lambda values. Default is 100.}

\item{maxit}{Maximum number of passes over the data for all lambda values. Default is \eqn{10^7}.}

\item{exact}{A logical flag indicating whether to use a sequence of lambda values (from \code{lambda_max} to \code{lambda_best})
when fitting the final model on the entire dataset.
The functions in the \code{glmnet} package are designed for efficiency by computing the entire regularization path
(a sequence of lambda values) using "warm starts", which is often faster than computing a single fit. The default is \code{TRUE}.}
}
\value{
A 'global' model which is icomb coherent
}
\description{
Information combination reconciliation
}
